requirements.txt
streamlit
patool
gunicorn
fastapi
uvicorn
loguru
python-multipart
psycopg2-binary
requests
pydantic
librosa
aiofiles
python-dotenv
httpx
rarfile